Changed processing order of patch_form_preprocessor and patch_columns.
Patch columns should come first so that we can change some attributes in
preprocessor and we don't have to declare them in patch_columns.
Attribute last_modified is a good example - we don't need to have it in
patch_columns as user should not be able to set this attribute but we
also want to be able to set it in patch_form_preprocessor.