Rewrote code from child collection access.
Previously the url_rule was in form "entity/<instid>/<relation>/". That
way it was not possible to enhance blueprint with other custom endpoints
related to the entity. Current solution creates url_rule in form
"entity/<instid>/relation/". Therefore relation is not a variable as
before but it is set explicitly from model introspection.
That allows creation of custom "entity/<instid>/custom_relation" rules
in the blueprint.